+++ generator_tests/test/_dummy.dart
@@ -184,12 +184,99 @@
 //       form.control(fruitListControlPath()) as FormArray<bool>;
 //   FormArray<String> get vegetablesListControl =>
 //       form.control(vegetablesListControlPath()) as FormArray<String>;
-//   void addEmailListItem(String value) =>
-//       emailListControl.add(FormControl<String>(value: value));
-//   void addFruitListItem(bool value) =>
-//       fruitListControl.add(FormControl<bool>(value: value));
-//   void addVegetablesListItem(String value) =>
-//       vegetablesListControl.add(FormControl<String>(value: value));
+//   void addEmailListItem(String value,
+//       {List<AsyncValidatorFunction>? asyncValidators,
+//       List<ValidatorFunction>? validators,
+//       int? asyncValidatorsDebounceTime,
+//       bool? disabled,
+//       ValidatorsApplyMode validatorsApplyMode = ValidatorsApplyMode.merge}) {
+//     List<ValidatorFunction> resultingValidators = [];
+//     List<AsyncValidatorFunction> resultingAsyncValidators = [];
+//
+//     switch (validatorsApplyMode) {
+//       case ValidatorsApplyMode.merge:
+//         if (validators != null) resultingValidators.addAll(validators);
+//         if (asyncValidators != null)
+//           resultingAsyncValidators.addAll(asyncValidators);
+//         break;
+//       case ValidatorsApplyMode.override:
+//         if (validators != null) resultingValidators = validators;
+//
+//         if (asyncValidators != null) resultingAsyncValidators = asyncValidators;
+//         break;
+//     }
+//
+//     emailListControl.add(FormControl<String>(
+//       value: value,
+//       validators: resultingValidators,
+//       asyncValidators: resultingAsyncValidators,
+//       asyncValidatorsDebounceTime: asyncValidatorsDebounceTime ?? 250,
+//       disabled: disabled ?? false,
+//     ));
+//   }
+//
+//   void addFruitListItem(bool value,
+//       {List<AsyncValidatorFunction>? asyncValidators,
+//       List<ValidatorFunction>? validators,
+//       int? asyncValidatorsDebounceTime,
+//       bool? disabled,
+//       ValidatorsApplyMode validatorsApplyMode = ValidatorsApplyMode.merge}) {
+//     List<ValidatorFunction> resultingValidators = [];
+//     List<AsyncValidatorFunction> resultingAsyncValidators = [];
+//
+//     switch (validatorsApplyMode) {
+//       case ValidatorsApplyMode.merge:
+//         if (validators != null) resultingValidators.addAll(validators);
+//         if (asyncValidators != null)
+//           resultingAsyncValidators.addAll(asyncValidators);
+//         break;
+//       case ValidatorsApplyMode.override:
+//         if (validators != null) resultingValidators = validators;
+//
+//         if (asyncValidators != null) resultingAsyncValidators = asyncValidators;
+//         break;
+//     }
+//
+//     fruitListControl.add(FormControl<bool>(
+//       value: value,
+//       validators: resultingValidators,
+//       asyncValidators: resultingAsyncValidators,
+//       asyncValidatorsDebounceTime: asyncValidatorsDebounceTime ?? 250,
+//       disabled: disabled ?? false,
+//     ));
+//   }
+//
+//   void addVegetablesListItem(String value,
+//       {List<AsyncValidatorFunction>? asyncValidators,
+//       List<ValidatorFunction>? validators,
+//       int? asyncValidatorsDebounceTime,
+//       bool? disabled,
+//       ValidatorsApplyMode validatorsApplyMode = ValidatorsApplyMode.merge}) {
+//     List<ValidatorFunction> resultingValidators = [];
+//     List<AsyncValidatorFunction> resultingAsyncValidators = [];
+//
+//     switch (validatorsApplyMode) {
+//       case ValidatorsApplyMode.merge:
+//         if (validators != null) resultingValidators.addAll(validators);
+//         if (asyncValidators != null)
+//           resultingAsyncValidators.addAll(asyncValidators);
+//         break;
+//       case ValidatorsApplyMode.override:
+//         if (validators != null) resultingValidators = validators;
+//
+//         if (asyncValidators != null) resultingAsyncValidators = asyncValidators;
+//         break;
+//     }
+//
+//     vegetablesListControl.add(FormControl<String>(
+//       value: value,
+//       validators: resultingValidators,
+//       asyncValidators: resultingAsyncValidators,
+//       asyncValidatorsDebounceTime: asyncValidatorsDebounceTime ?? 250,
+//       disabled: disabled ?? false,
+//     ));
+//   }
+//
 //   ArrayNullable get model => ArrayNullable(
 //       emailList: emailListValue,
 //       fruitList: fruitListValue,
@@ -200,7 +287,13 @@
 //   FormGroup formElements() => FormGroup({
 //         emailListControlName: FormArray<String>(
 //             arrayNullable.emailList
-//                 .map((e) => FormControl<String>(value: e))
+//                 .map((e) => FormControl<String>(
+//                       value: e,
+//                       validators: [],
+//                       asyncValidators: [],
+//                       asyncValidatorsDebounceTime: 250,
+//                       disabled: false,
+//                     ))
 //                 .toList(),
 //             validators: [],
 //             asyncValidators: [],
@@ -208,7 +301,13 @@
 //             disabled: false),
 //         fruitListControlName: FormArray<bool>(
 //             arrayNullable.fruitList
-//                 .map((e) => FormControl<bool>(value: e))
+//                 .map((e) => FormControl<bool>(
+//                       value: e,
+//                       validators: [],
+//                       asyncValidators: [],
+//                       asyncValidatorsDebounceTime: 250,
+//                       disabled: false,
+//                     ))
 //                 .toList(),
 //             validators: [],
 //             asyncValidators: [],
@@ -216,7 +315,13 @@
 //             disabled: false),
 //         vegetablesListControlName: FormArray<String>(
 //             arrayNullable.vegetablesList
-//                     ?.map((e) => FormControl<String>(value: e))
+//                     ?.map((e) => FormControl<String>(
+//                           value: e,
+//                           validators: [],
+//                           asyncValidators: [],
+//                           asyncValidatorsDebounceTime: 250,
+//                           disabled: false,
+//                         ))
 //                     .toList() ??
 //                 [],
 //             validators: [],
@@ -236,4 +341,3 @@
 //           asyncValidatorsDebounceTime: 250,
 //           disabled: false);
 // }
-// '