+++ packages/reactive_forms_annotations/lib/src/form_array_annotation.dart
@@ -1,5 +1,10 @@
 import 'package:reactive_forms/reactive_forms.dart';
 
+enum ValidatorsApplyMode {
+  merge,
+  override,
+}
+
 class FormArrayAnnotation {
   final List<ValidatorFunction> validators;
 
@@ -9,10 +14,22 @@ class FormArrayAnnotation {
 
   final bool disabled;
 
+  final List<ValidatorFunction> itemValidators;
+
+  final List<AsyncValidatorFunction> itemAsyncValidators;
+
+  final int itemAsyncValidatorsDebounceTime;
+
+  final bool itemDisabled;
+
   const FormArrayAnnotation({
     this.validators = const [],
     this.asyncValidators = const [],
     this.asyncValidatorsDebounceTime = 250,
     this.disabled = false,
+    this.itemValidators = const [],
+    this.itemAsyncValidators = const [],
+    this.itemAsyncValidatorsDebounceTime = 250,
+    this.itemDisabled = false,
   });
 }