+++ packages/reactive_forms_generator/example/lib/docs/arrays/mailing_list.dart
@@ -7,6 +7,9 @@ class MailingList {
     validators: const [
       mailingListValidator,
     ],
+    itemValidators: [
+      emailValidator,
+    ],
   )
   final List<String?> emailList;