+++ packages/reactive_forms_generator/example/lib/routes/array_nullable_form.dart
@@ -103,10 +103,10 @@ class ArrayNullableFormWidget extends StatelessWidget {
                 child: const Text('Sign Up'),
               ),
               ReactiveArrayNullableFormConsumer(
-                builder: (context, form, child) {
+                builder: (context, formModel, child) {
                   return ElevatedButton(
                     child: Text('Submit'),
-                    onPressed: form.form.valid ? () {} : null,
+                    onPressed: formModel.form.valid ? () {} : null,
                   );
                 },
               ),