+++ packages/reactive_forms_generator/example/lib/routes/login_form.dart
@@ -129,10 +129,10 @@ class LoginFormWidget extends StatelessWidget {
                 child: const Text('Sign Up'),
               ),
               ReactiveLoginFormConsumer(
-                builder: (context, form, child) {
+                builder: (context, formModel, child) {
                   return ElevatedButton(
                     child: Text('Submit'),
-                    onPressed: form.form.valid ? () {} : null,
+                    onPressed: formModel.form.valid ? () {} : null,
                   );
                 },
               ),