+++ packages/reactive_forms_generator/example/lib/routes/login_nullable_form.dart
@@ -121,10 +121,10 @@ class LoginNullableFormWidget extends StatelessWidget {
                 child: const Text('Sign Up'),
               ),
               ReactiveLoginNullableFormConsumer(
-                builder: (context, form, child) {
+                builder: (context, formModel, child) {
                   return ElevatedButton(
                     child: Text('Submit'),
-                    onPressed: form.form.valid ? () {} : null,
+                    onPressed: formModel.form.valid ? () {} : null,
                   );
                 },
               ),