+++ packages/reactive_forms_annotations/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.5-beta]
+
+* extended FormArray annotation
+
 ## [0.0.4-beta]
 
 * asyncDebounceTime and disable support